backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
mangum==0.17.0
python-dotenv==1.0.1
python-multipart==0.0.9
boto3==1.35.36
